Update patch set 1
Patch Set 1:

An important note is that for the build to be successful with these changes, the lunch target omni_sailfish needs to be used (refer #1769). When I attempted to use just this change on the lineage_sailfish target, it failed to boot properly - blank screen and reboot. So if this is accepted, any release versions will have to be built using omni_sailfish.

Patch-set: 1